+++ packages/contracts/src/systems/_RegistryCreateQuestObjectiveSystem.sol
@@ -17,13 +17,13 @@ contract _RegistryCreateQuestObjectiveSystem is System {
 
   function execute(bytes memory arguments) public onlyOwner returns (bytes memory) {
     (
-      uint256 questIndex,
+      uint32 questIndex,
       string memory name,
       string memory logicType,
       string memory type_,
-      uint256 index, // generic index
+      uint32 index, // generic index
       uint256 value
-    ) = abi.decode(arguments, (uint256, string, string, string, uint256, uint256));
+    ) = abi.decode(arguments, (uint32, string, string, string, uint32, uint256));
 
     // check that the quest exists
     uint256 questID = LibRegistryQuests.getByQuestIndex(components, questIndex);
@@ -46,11 +46,11 @@ contract _RegistryCreateQuestObjectiveSystem is System {
   }
 
   function executeTyped(
-    uint256 questIndex,
+    uint32 questIndex,
     string memory name,
     string memory logicType,
     string memory type_,
-    uint256 index, // can be empty
+    uint32 index, // can be empty
     uint256 value
   ) public onlyOwner returns (bytes memory) {
     return execute(abi.encode(questIndex, name, logicType, type_, index, value));
